fix(cli): preserve model/model-spec in heterogeneous tool selection
Route HeterogeneousStrict through resolve_tool_and_model() after
selecting the heterogeneous tool, so --model and --model-spec flags
are properly resolved instead of silently dropped.

Also extract execute_with_session and determine_project_root from
main.rs to pipeline.rs to stay under the 800-line monolith limit.
main.rs: 806 -> 586 lines.
